test: convert analytics-sources tests to jest
